+++ backend/controllers/index.js
@@ -1,0 +1,11 @@
+const remedyController = require('./remedyController');
+const recommendationsController = require('./recommendationsController');
+const symptomsController = require('./symptomsController');
+const userController = require('./userController');
+
+module.exports = {
+    remedyController: remedyController,
+    recommendationsController: recommendationsController,
+    userController: userController,
+    symptomsController: symptomsController
+}
