+++ backend/routes/index.js
@@ -1,13 +1,13 @@
 const auth = require('./auth');
-const getRemedyRecommendation = require('./getRemedyRecommendation');
-const getSymptoms = require('./getSymptoms');
+const recommendations = require('./recommendations');
+const symptoms = require('./symptoms');
 const remedies = require('./remedies');
 const user = require('./user');
 
 module.exports = {
     user: user,
     remedies: remedies,
-    getSymptoms: getSymptoms,
-    getRemedyRecommendation: getRemedyRecommendation,
+    symptoms: symptoms,
+    recommendations: recommendations,
     auth: auth
 }
